third_party/WebKit/LayoutTests/fast/dom/shadow/pseudo-not-traversal.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>

<template id="item">
    <div id="host1">
        <span class="a">First</span>
        <span class="b">Second</span>
    </div>
</template>

<template id="host1-shadow">
    <div id="host2">
        <content select=".b"></content>
        <content select=".a"></content>
    </div>
</template>

<template id="host2-shadow">
    <content id="target"></content>
</template>

<script>
// Creates the host/shadow templates described above which swap the nodes
// using distribution and then returns a string list of the classes of the
// elements that were distributed in order.
function testSelector(select) {
    var host1 = document.getElementById("item").content.firstElementChild.cloneNode(true);
    var template1 = document.getElementById("host1-shadow").content;
    host1.createShadowRoot().appendChild(template1.cloneNode(true));

    var host2 = host1.shadowRoot.getElementById("host2");
    var template2 = document.getElementById("host2-shadow").content;
    host2.createShadowRoot().appendChild(template2.cloneNode(true));

    var target = host2.shadowRoot.getElementById("target");
    target.select = select;

    var result = [];
    var nodes = target.getDistributedNodes();
    for (var i = 0; i < nodes.length; ++i)
        result.push(nodes[i].className);

    return result.join(",");
}

description("Test the :not pseudo selector in <content select>");

shouldBeEqualToString("testSelector('span')", "b,a");
shouldBeEqualToString("testSelector(':nth-child(1)')", "b");
shouldBeEqualToString("testSelector(':nth-child(2)')", "a");
shouldBeEqualToString("testSelector(':not(:nth-child(1))')", "a");
shouldBeEqualToString("testSelector(':not(:nth-child(2))')", "b");
</script>
