third_party/WebKit/LayoutTests/fast/dom/shadow/pseudo-not-traversal-expected.txt

Test the :not pseudo selector in

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSelector('span') is "b,a"
PASS testSelector(':nth-child(1)') is "b"
PASS testSelector(':nth-child(2)') is "a"
PASS testSelector(':not(:nth-child(1))') is "a"
PASS testSelector(':not(:nth-child(2))') is "b"
PASS successfullyParsed is true

TEST COMPLETE
